+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1 (unreleased)
 - fix: default object
 - better EML files loading
+- generic header manipulation (removing, adding multiple, mostly preserves order when loading)
 
 ## 1.0.0 (2020-05-28)
 - CHANGED: